sdk/python/src/dataflare.py
import yaml
from pyecharts import options as opts
from pyecharts.charts import Line, Bar

def render_chart_from_datacard(datacard, div_id):
    title = datacard['title']
    vis_type = datacard['visualization']['type']
    
    x_axis = datacard['visualization']['x_axis']
    y_axis = datacard['visualization']['y_axis']
    
    # Mock data for the example. Replace with API or CSV data as in datacard.
    x_data = ["2021-01", "2021-02", "2021-03"]
    y_data1 = [10, 20, 30]
    y_data2 = [5, 15, 25]
    
    if vis_type == "chart":
        chart_type = datacard['visualization']['chart_type']
        
        if chart_type == "line":
            chart = Line()
        elif chart_type == "bar":
            chart = Bar()
            
        chart.add_xaxis(x_data)
        
        for i, y in enumerate(y_axis):
            chart.add_yaxis(y, y_data1 if i == 0 else y_data2)
            
        chart.set_global_opts(title_opts=opts.TitleOpts(title=title))
        chart.render_notebook()
        
        # If you want to render to a specific div in an HTML file
        chart.render_embed(div_id)
        